Turns out that using std::vector as the underlying storage for images has a performance hit when resizing. I think this is due to the assurances that vector constructs and destructs elements correctly, but this has an overhead even when using fundamental types for some reason. Since we assume image data is invalid when resizing we don't care about copying data and images should always be fundamental types so it is considerably more efficient to continue using a custom buffer. But the struct/Buffer class should still be deprecated and std::vector used in most other places, so have created a new Buffer class with just the bare essentials that is specific to image storage. It also tries to meet the requirements of Container as much as possible, so it can be used most places that std::vector can be used.
Implemented unit tests for the new Buffer structure, hopefully these are reasonably comprehensive.

Added a contiguity test for container types - but it will only work properly in C++ when it can
check against `contiguous_iterator_tag`.

Deprecated Operations and replaced the corresponding calls in ImageBase with STL equivalents where
possible.

Implemented Row and Column iterators for SubImage and added a function to generated a sub-image
from ImageBase<>. Updated the Inspect() function to use a sub-image.