An attempt at a couple of event publishers. KeyPublisher allows subscription to events based on a key (which could be a string or enum for example). The event type is fixed. PolymorphicPublisher allows subscription to events based on the event type - this permits subscription to the base event type which will thereby receive all events.
Both require the subscriptions to be polled so that the callbacks are invoked on the same thread as the listener. This means
that events are placed in a queue so a lack of polling will allow the queue to fill and then start dropping events.

The subscription object will automatically detach from the publisher when destroyed and so it must not be permitted to
live beyond the lifetime of the publisher.